+++ Demos/AllOverIt.ReactiveUI/ViewRegistryDemo/ViewModels/View2ViewModel.cs
@@ -1,0 +1,17 @@
+﻿using AllOverIt.ReactiveUI;
+using ReactiveUI.Fody.Helpers;
+using System.Reactive.Disposables;
+
+namespace ViewRegistryDemo.ViewModels
+{
+    public sealed class View2ViewModel : ActivatableViewModel
+    {
+        [Reactive]
+        public int Id { get; set; }
+
+        protected override void OnActivated(CompositeDisposable disposables)
+        {
+
+        }
+    }
+}