+++ Demos/AllOverIt.ReactiveUI/ViewRegistryDemo/ViewRegistry/IDemoViewRegistry.cs
@@ -1,0 +1,10 @@
+﻿using AllOverIt.ReactiveUI.ViewRegistry;
+
+namespace ViewRegistryDemo.ViewRegistry
+{
+    // The 'int' is the Id type for each view in the registry
+    public interface IDemoViewRegistry : IViewRegistry<int>
+    {
+
+    }
+}